chore(deezer-sdk): split files out from index
